refactor: grpc logic to handler
