+++ jina/serve/runtimes/worker/request_handling.py
@@ -1,17 +1,30 @@
-import asyncio
 import functools
 import json
-from typing import TYPE_CHECKING, Dict, List, Optional, Tuple
+import os
+from typing import Dict, Tuple
 
 from jina._docarray import DocumentArray
 from jina.constants import __default_endpoint__
 from jina.excepts import BadConfigSource
-from jina.importer import ImportExtensions
 from jina.serve.executors import BaseExecutor
 from jina.serve.runtimes.worker.batch_queue import BatchQueue
+import argparse
+import asyncio
+from typing import TYPE_CHECKING, AsyncIterator, List, Optional
+
+import grpc
+
+from jina.excepts import RuntimeTerminated
+from jina.helper import get_full_version
+from jina.importer import ImportExtensions
+from jina.proto import jina_pb2
+from jina.serve.instrumentation import MetricsTimer
 from jina.types.request.data import DataRequest
 
 if TYPE_CHECKING:  # pragma: no cover
+    from opentelemetry.propagate import Context
+
+    from jina.types.request import Request
     import argparse
 
     from opentelemetry import metrics, trace
@@ -27,14 +40,16 @@ class WorkerRequestHandler:
     _KEY_RESULT = '__results__'
 
     def __init__(
-        self,
-        args: 'argparse.Namespace',
-        logger: 'JinaLogger',
-        metrics_registry: Optional['CollectorRegistry'] = None,
-        tracer_provider: Optional['trace.TracerProvider'] = None,
-        meter_provider: Optional['metrics.MeterProvider'] = None,
-        deployment_name: str = '',
-        **kwargs,
+            self,
+            args: 'argparse.Namespace',
+            logger: 'JinaLogger',
+            metrics_registry: Optional['CollectorRegistry'] = None,
+            tracer_provider: Optional['trace.TracerProvider'] = None,
+            meter_provider: Optional['metrics.MeterProvider'] = None,
+            meter=None,
+            tracer=None,
+            deployment_name: str = '',
+            **kwargs,
     ):
         """Initialize private parameters and execute private loading functions.
 
@@ -47,9 +62,67 @@ class WorkerRequestHandler:
         :param kwargs: extra keyword arguments
         """
         super().__init__()
+        self.meter = meter
+        self.metrics_registry = metrics_registry
+        self.tracer = tracer
         self.args = args
         self.logger = logger
         self._is_closed = False
+        if self.metrics_registry:
+            with ImportExtensions(
+                    required=True,
+                    help_text='You need to install the `prometheus_client` to use the montitoring functionality of jina',
+            ):
+                from prometheus_client import Counter, Summary
+
+            self._summary = Summary(
+                'receiving_request_seconds',
+                'Time spent processing request',
+                registry=self.metrics_registry,
+                namespace='jina',
+                labelnames=('runtime_name',),
+            ).labels(self.args.name)
+
+            self._failed_requests_metrics = Counter(
+                'failed_requests',
+                'Number of failed requests',
+                registry=self.metrics_registry,
+                namespace='jina',
+                labelnames=('runtime_name',),
+            ).labels(self.args.name)
+
+            self._successful_requests_metrics = Counter(
+                'successful_requests',
+                'Number of successful requests',
+                registry=self.metrics_registry,
+                namespace='jina',
+                labelnames=('runtime_name',),
+            ).labels(self.args.name)
+
+        else:
+            self._summary = None
+            self._failed_requests_metrics = None
+            self._successful_requests_metrics = None
+
+        if self.meter:
+            self._receiving_request_seconds = self.meter.create_histogram(
+                name='jina_receiving_request_seconds',
+                description='Time spent processing request',
+            )
+            self._failed_requests_counter = self.meter.create_counter(
+                name='jina_failed_requests',
+                description='Number of failed requests',
+            )
+
+            self._successful_requests_counter = self.meter.create_counter(
+                name='jina_successful_requests',
+                description='Number of successful requests',
+            )
+        else:
+            self._receiving_request_seconds = None
+            self._failed_requests_counter = None
+            self._successful_requests_counter = None
+        self._metric_attributes = {'runtime_name': self.args.name}
         self._load_executor(
             metrics_registry=metrics_registry,
             tracer_provider=tracer_provider,
@@ -68,6 +141,39 @@ class WorkerRequestHandler:
         # the below is of "shape" exec_endpoint_name -> parameters_key -> batch_queue
         self._batchqueue_instances: Dict[str, Dict[str, BatchQueue]] = {}
         self._init_batchqueue_dict()
+        self._hot_reload_task = None
+        if self.args.reload:
+            self._hot_reload_task = asyncio.create_task(self._hot_reload())
+
+    async def _hot_reload(self):
+        import inspect
+
+        executor_file = inspect.getfile(self._executor.__class__)
+        watched_files = set([executor_file] + (self.args.py_modules or []))
+        executor_base_path = os.path.dirname(os.path.abspath(executor_file))
+        extra_paths = [
+            os.path.join(path, name)
+            for path, subdirs, files in os.walk(executor_base_path)
+            for name in files
+        ]
+        extra_python_paths = list(filter(lambda x: x.endswith('.py'), extra_paths))
+        for extra_python_file in extra_python_paths:
+            watched_files.add(extra_python_file)
+
+        with ImportExtensions(
+                required=True,
+                logger=self.logger,
+                help_text='''hot reload requires watchfiles dependency to be installed. You can do `pip install 
+                watchfiles''',
+        ):
+            from watchfiles import awatch
+
+        async for changes in awatch(*watched_files):
+            changed_files = [changed_file for _, changed_file in changes]
+            self.logger.info(
+                f'detected changes in: {changed_files}. Refreshing the Executor'
+            )
+            self._refresh_executor(changed_files)
 
     def _all_batch_queues(self) -> List[BatchQueue]:
         """Returns a list of all batch queue instances
@@ -123,16 +229,16 @@ class WorkerRequestHandler:
             }
 
     def _init_monitoring(
-        self,
-        metrics_registry: Optional['CollectorRegistry'] = None,
-        meter: Optional['metrics.Meter'] = None,
+            self,
+            metrics_registry: Optional['CollectorRegistry'] = None,
+            meter: Optional['metrics.Meter'] = None,
     ):
 
         if metrics_registry:
 
             with ImportExtensions(
-                required=True,
-                help_text='You need to install the `prometheus_client` to use the montitoring functionality of jina',
+                    required=True,
+                    help_text='You need to install the `prometheus_client` to use the montitoring functionality of jina',
             ):
                 from prometheus_client import Counter, Summary
 
@@ -188,10 +294,10 @@ class WorkerRequestHandler:
             self._sent_response_size_histogram = None
 
     def _load_executor(
-        self,
-        metrics_registry: Optional['CollectorRegistry'] = None,
-        tracer_provider: Optional['trace.TracerProvider'] = None,
-        meter_provider: Optional['metrics.MeterProvider'] = None,
+            self,
+            metrics_registry: Optional['CollectorRegistry'] = None,
+            tracer_provider: Optional['trace.TracerProvider'] = None,
+            meter_provider: Optional['metrics.MeterProvider'] = None,
     ):
         """
         Load the executor to this runtime, specified by ``uses`` CLI argument.
@@ -375,7 +481,7 @@ class WorkerRequestHandler:
         return docs
 
     async def handle(
-        self, requests: List['DataRequest'], tracing_context: Optional['Context'] = None
+            self, requests: List['DataRequest'], tracing_context: Optional['Context'] = None
     ) -> DataRequest:
         """Initialize private parameters and execute private loading functions.
 
@@ -452,7 +558,7 @@ class WorkerRequestHandler:
 
     @staticmethod
     def replace_docs(
-        request: List['DataRequest'], docs: 'DocumentArray', ndarray_type: str = None
+            request: List['DataRequest'], docs: 'DocumentArray', ndarray_type: str = None
     ) -> None:
         """Replaces the docs in a message with new Documents.
 
@@ -488,6 +594,8 @@ class WorkerRequestHandler:
 
     async def close(self):
         """Close the data request handler, by closing the executor and the batch queues."""
+        if self._hot_reload_task is not None:
+            self._hot_reload_task.cancel()
         if not self._is_closed:
             await asyncio.gather(*[q.close() for q in self._all_batch_queues()])
             self._executor.close()
@@ -495,7 +603,7 @@ class WorkerRequestHandler:
 
     @staticmethod
     def _get_docs_matrix_from_request(
-        requests: List['DataRequest'],
+            requests: List['DataRequest'],
     ) -> Tuple[Optional[List['DocumentArray']], Optional[Dict[str, 'DocumentArray']]]:
         """
         Returns a docs matrix from a list of DataRequest objects.
@@ -519,7 +627,7 @@ class WorkerRequestHandler:
 
     @staticmethod
     def get_parameters_dict_from_request(
-        requests: List['DataRequest'],
+            requests: List['DataRequest'],
     ) -> 'Dict':
         """
         Returns a parameters dict from a list of DataRequest objects.
@@ -539,7 +647,7 @@ class WorkerRequestHandler:
 
     @staticmethod
     def get_docs_from_request(
-        requests: List['DataRequest'],
+            requests: List['DataRequest'],
     ) -> 'DocumentArray':
         """
         Gets a field from the message
@@ -606,3 +714,130 @@ class WorkerRequestHandler:
         WorkerRequestHandler.replace_parameters(requests[0], params)
 
         return requests[0]
+
+    # serving part
+    async def process_single_data(self, request: DataRequest, context) -> DataRequest:
+        """
+        Process the received requests and return the result as a new request
+
+        :param request: the data request to process
+        :param context: grpc context
+        :returns: the response request
+        """
+
+        return await self.process_data([request], context)
+
+    async def endpoint_discovery(self, empty, context) -> jina_pb2.EndpointsProto:
+        """
+        Process the the call requested and return the list of Endpoints exposed by the Executor wrapped inside this Runtime
+
+        :param empty: The service expects an empty protobuf message
+        :param context: grpc context
+        :returns: the response request
+        """
+        self.logger.debug('got an endpoint discovery request')
+        endpoints_proto = jina_pb2.EndpointsProto()
+        endpoints_proto.endpoints.extend(
+            list(self._executor.requests.keys())
+        )
+        return endpoints_proto
+
+    def _extract_tracing_context(
+            self, metadata: grpc.aio.Metadata
+    ) -> Optional['Context']:
+        if self.tracer:
+            from opentelemetry.propagate import extract
+
+            context = extract(dict(metadata))
+            return context
+
+        return None
+
+    async def process_data(self, requests: List[DataRequest], context) -> DataRequest:
+        """
+        Process the received requests and return the result as a new request
+
+        :param requests: the data requests to process
+        :param context: grpc context
+        :returns: the response request
+        """
+        with MetricsTimer(
+                self._summary, self._receiving_request_seconds, self._metric_attributes
+        ):
+            try:
+                if self.logger.debug_enabled:
+                    self._log_data_request(requests[0])
+
+                tracing_context = self._extract_tracing_context(
+                    context.invocation_metadata()
+                )
+                result = await self.handle(
+                    requests=requests, tracing_context=tracing_context
+                )
+                if self._successful_requests_metrics:
+                    self._successful_requests_metrics.inc()
+                if self._successful_requests_counter:
+                    self._successful_requests_counter.add(
+                        1, attributes=self._metric_attributes
+                    )
+                return result
+            except (RuntimeError, Exception) as ex:
+                self.logger.error(
+                    f'{ex!r}'
+                    + f'\n add "--quiet-error" to suppress the exception details'
+                    if not self.args.quiet_error
+                    else '',
+                    exc_info=not self.args.quiet_error,
+                )
+
+                requests[0].add_exception(ex, self._executor)
+                context.set_trailing_metadata((('is-error', 'true'),))
+                if self._failed_requests_metrics:
+                    self._failed_requests_metrics.inc()
+                if self._failed_requests_counter:
+                    self._failed_requests_counter.add(
+                        1, attributes=self._metric_attributes
+                    )
+
+                if (
+                        self.args.exit_on_exceptions
+                        and type(ex).__name__ in self.args.exit_on_exceptions
+                ):
+                    self.logger.info('Exiting because of "--exit-on-exceptions".')
+                    raise RuntimeTerminated
+
+                return requests[0]
+
+    async def _status(self, empty, context) -> jina_pb2.JinaInfoProto:
+        """
+        Process the the call requested and return the JinaInfo of the Runtime
+
+        :param empty: The service expects an empty protobuf message
+        :param context: grpc context
+        :returns: the response request
+        """
+        self.logger.debug('recv _status request')
+        info_proto = jina_pb2.JinaInfoProto()
+        version, env_info = get_full_version()
+        for k, v in version.items():
+            info_proto.jina[k] = str(v)
+        for k, v in env_info.items():
+            info_proto.envs[k] = str(v)
+        return info_proto
+
+    async def stream(
+            self, request_iterator, context=None, *args, **kwargs
+    ) -> AsyncIterator['Request']:
+        """
+        stream requests from client iterator and stream responses back.
+
+        :param request_iterator: iterator of requests
+        :param context: context of the grpc call
+        :param args: positional arguments
+        :param kwargs: keyword arguments
+        :yield: responses to the request
+        """
+        async for request in request_iterator:
+            yield await self.process_data([request], context)
+
+    Call = stream
